+++ src/Cvoya.Spring.Host.Api/Program.cs
@@ -3,6 +3,7 @@
 
 using System.Text.Json.Serialization;
 
+using Cvoya.Spring.Connector.Arxiv.DependencyInjection;
 using Cvoya.Spring.Connector.GitHub.DependencyInjection;
 using Cvoya.Spring.Dapr.Auth;
 using Cvoya.Spring.Dapr.DependencyInjection;
@@ -28,6 +29,7 @@ builder.Services
     .AddCvoyaSpringDapr(builder.Configuration)
     .AddCvoyaSpringOllamaLlm(builder.Configuration)
     .AddCvoyaSpringConnectorGitHub(builder.Configuration)
+    .AddCvoyaSpringConnectorArxiv(builder.Configuration)
     .AddCvoyaSpringApiServices(builder.Configuration);
 
 // DataProtection tries to persist/load keys from disk and logs a warning when